Create `saveProjectToFirestore` util file
Related to #60
