+++ src/app/utils/firebase/index.ts
@@ -1,0 +1,5 @@
+import { saveProjectToFirestore } from "./saveProjectToFirestore";
+
+export {
+  saveProjectToFirestore
+}
